Relocate catalyst UI components to subfolder
