+++ components/compounds/InfoCard.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image'
 import { FlavorText, VersionGroupFlavorText } from 'pokedex-promise-v2'
 import GlassCard from '@/components/GlassCard'
 import { useVersionGroup } from '@/lib/stores/version-group'
-import { Badge } from '@/components/ui/badge'
+import { Badge } from '@/components/ui/catalyst/badge'
 
 type Tag = {
   label: string
