+++ components/compounds/MonsterCard.tsx
@@ -2,7 +2,7 @@
 
 import Image from 'next/image'
 import { motion } from 'motion/react'
-import Link from '@/components/ui/link'
+import Link from '@/components/ui/catalyst/link'
 import { memo } from 'react'
 import { TypeKey } from '@/lib/utils/pokeapi-helpers'
 import GlassCard from '@/components/GlassCard'
