+++ components/details/MonsterMetadata.tsx
@@ -1,4 +1,4 @@
-import Link from '@/components/ui/link'
+import Link from '@/components/ui/catalyst/link'
 import clsx from 'clsx/lite'
 import pMap from 'p-map'
 import type {
