+++ components/ui/catalyst/button.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from 'react'
 import clsx from 'clsx'
 import * as Headless from '@headlessui/react'
-import Link from '@/components/ui/link'
+import Link from '@/components/ui/catalyst/link'
 
 const styles = {
   base: [
