+++ components/ui/catalyst/table.tsx
@@ -3,7 +3,7 @@
 import type React from 'react'
 import { createContext, useContext, useState } from 'react'
 import clsx from 'clsx/lite'
-import Link from '@/components/ui/link'
+import Link from '@/components/ui/catalyst/link'
 
 const TableContext = createContext<{
   bleed: boolean
